+++ test_example.sh
@@ -13,17 +13,18 @@ cd $REPO
 git init
 
 # add remotes
-git remote add aws_ec2_describeInstances aws+ec2:///describe-instances
-git remote add aws_ec2_catalog aws+ec2:///describe-instances
+git remote add aws_ec2_describeInstances aws+ec2::/describe-instances
+git remote add aws_ec2_catalog aws+ec2::/catalog
+git remote -v
 
 # fetch
 git fetch aws_ec2_describeInstances
 git fetch aws_ec2_catalog
-tree
 
 # force an error
-git remote add wrong aws+ec2://blabla/describe-instances
-git fetch wrong
+#git remote add wrong aws+ec2://blabla/catalog
+#git fetch wrong
 
 #-------------------
+tree
 echo "Test complete"